requirements.txt
Flask
Flask-Limiter
python-dotenv
transformers
torch
nltk
numpy
pdfminer.six
python-docx
openpyxl
python-pptx
Pillow
pytesseract
cachetools
markdown
reportlab
wordcloud
matplotlib
gunicorn
accelerate
Flask-Cors
Flask-JWT-Extended
marshmallow
pytest
black
flake8
